added lookup of shared boards and limited deletion of board to board owner

+++ controllers/boardController.js
@@ -4,8 +4,13 @@ const List = require("../models/listModel");
 
 const getBoards = async (req, res) => {
   const boards = await Board.find({ user: req.user.id });
-
-  res.status(200).json(boards);
+  const sharedBoards = await Board.find({
+    assignedUsers: { $elemMatch: { userId: req.user.id } },
+  });
+  res.status(200).json({
+    boards,
+    sharedBoards,
+  });
 };
 
 const setBoards = async (req, res) => {
@@ -18,16 +23,23 @@ const setBoards = async (req, res) => {
   });
   res.status(200).json(board);
 };
+
 const updateBoards = (req, res) => {
   res.status(200).send({ title: "updatedTitle" });
 };
+
 const deleteBoard = async (req, res) => {
   const id = req.params.id;
-
-  const deletedBoard = await Board.findOneAndDelete({ _id: id });
-  const deleteBoardList = await List.deleteMany({ board: id });
-  const boards = await Board.find({ user: req.user.id });
+  const deletedBoard = await Board.findOneAndDelete({ _id: id, user: req.user._id});
+  if(deletedBoard !== null){
+    const deleteBoardList = await List.deleteMany({ board: id });
+    const boards = await Board.find({ user: req.user.id });
   res.status(200).json(boards);
+  }else{
+    res.status(400).json(
+      "Not authorized to delete this board"
+    )
+  }
 };
 
 module.exports = { getBoards, setBoards, updateBoards, deleteBoard };
